crate files for widgets and add opening link for build route
