+++ lib/src/features/onboarding_performance/view/widgets/page_indicator.dart
@@ -1,9 +1,42 @@
 import 'package:flutter/material.dart';
+import 'package:shebalin/src/features/onboarding_performance/models/onboard_performance.dart';
+
+class PageIndicator extends StatelessWidget {
+  const PageIndicator({
+    super.key,
+    required this.pages,
+    required this.currentIndex,
+  });
+
+  final List<OnboardPerformance> pages;
+  final int currentIndex;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      mainAxisAlignment: MainAxisAlignment.spaceAround,
+      children: [
+        ...List.generate(
+          pages.length,
+          (index) => DotIndicator(
+            count: pages.length,
+            isActive: index == currentIndex,
+          ),
+        ),
+      ],
+    );
+  }
+}
 
 class DotIndicator extends StatelessWidget {
   final int count;
   final bool isActive;
-  const DotIndicator({super.key, required this.count, required this.isActive});
+  const DotIndicator({
+    super.key,
+    required this.count,
+    required this.isActive,
+  });
 
   @override
   Widget build(BuildContext context) {
